requirements.txt
pillow
flask
numpy